Move search page under screens folder
Refs #366.
